feat(web): shared AppBanner host for top-of-app banners
Extract the fixed-position banner shell — main-container offset tracking
(useMainContainerOffset) and per-device localStorage dismissal
(useBannerDismiss) — into one AppBanner host that renders an Opal
MessageCard. Migrate LicenseExpiryBanner onto it, dropping its
copy-pasted offset hook, positioning wrapper, and dismiss state. The
caller keeps the copy/variant/visibility; the host owns position +
dismiss. No behavior change.
