+++ web/src/sections/LicenseExpiryBanner.tsx
@@ -1,10 +1,8 @@
 "use client";
 
-import { useEffect, useState } from "react";
-import { usePathname } from "next/navigation";
-import { MessageCard } from "@opal/components";
 import type { ExpiryWarningStage } from "@/lib/billing/interfaces";
 import { useLicense } from "@/hooks/useLicense";
+import AppBanner from "@/sections/AppBanner";
 
 const DISMISS_STORAGE_KEY = "license-expiry-banner-dismissed";
 
@@ -74,127 +72,36 @@ function dismissKey(
 ): string {
   const base = `${DISMISS_STORAGE_KEY}:${stage}:${expiresAt ?? "unknown"}`;
   if (stage === "grace") {
+    // Re-show once per day during the grace period so the countdown stays seen.
     const today = new Date().toISOString().slice(0, 10);
     return `${base}:${today}`;
   }
   return base;
 }
 
-interface LicenseExpiryBannerViewProps {
-  stage: ExpiryWarningStage;
-  expiresAt: string | null;
-  graceDaysRemaining: number;
-  onDismiss?: () => void;
-}
-
-export function LicenseExpiryBannerView({
-  stage,
-  expiresAt,
-  graceDaysRemaining,
-  onDismiss,
-}: LicenseExpiryBannerViewProps) {
-  const copy = buildCopy(stage, expiresAt, graceDaysRemaining);
-  if (!copy) return null;
-
-  return (
-    <MessageCard
-      variant={copy.variant}
-      title={copy.title}
-      description={copy.description}
-      onClose={onDismiss}
-    />
-  );
-}
-
-function useMainContainerOffset(): { left: number; width: number } {
-  const pathname = usePathname();
-  const [bounds, setBounds] = useState<{ left: number; width: number }>({
-    left: 0,
-    width: 0,
-  });
-
-  useEffect(() => {
-    if (typeof window === "undefined") return;
-    let target: HTMLElement | null = null;
-    let frame = 0;
-
-    function update() {
-      const el = document.querySelector<HTMLElement>("[data-main-container]");
-      if (el) {
-        const rect = el.getBoundingClientRect();
-        setBounds({ left: rect.left, width: rect.width });
-        if (target !== el) {
-          ro.disconnect();
-          ro.observe(el);
-          target = el;
-        }
-      } else {
-        setBounds({ left: 0, width: window.innerWidth });
-        target = null;
-      }
-    }
-
-    const ro = new ResizeObserver(update);
-    const mo = new MutationObserver(() => {
-      cancelAnimationFrame(frame);
-      frame = requestAnimationFrame(update);
-    });
-
-    update();
-    mo.observe(document.body, { childList: true, subtree: true });
-    window.addEventListener("resize", update);
-
-    return () => {
-      cancelAnimationFrame(frame);
-      ro.disconnect();
-      mo.disconnect();
-      window.removeEventListener("resize", update);
-    };
-  }, [pathname]);
-
-  return bounds;
-}
-
 export default function LicenseExpiryBanner() {
   const { data } = useLicense();
-  const [dismissed, setDismissed] = useState(false);
-  const { left, width } = useMainContainerOffset();
 
   const stage = data?.expiry_warning_stage ?? "none";
   const expiresAt = data?.expires_at ?? null;
   const graceDays = computeGraceDaysRemaining(data?.grace_period_end ?? null);
   const hasLicense = data?.has_license ?? false;
-  const key = dismissKey(stage, expiresAt);
 
-  useEffect(() => {
-    if (typeof window === "undefined") return;
-    setDismissed(window.localStorage.getItem(key) === "1");
-  }, [key]);
-
-  if (!hasLicense || stage === "none" || dismissed) {
+  if (!hasLicense || stage === "none") {
     return null;
   }
 
-  function handleDismiss() {
-    if (typeof window !== "undefined") {
-      window.localStorage.setItem(key, "1");
-    }
-    setDismissed(true);
+  const copy = buildCopy(stage, expiresAt, graceDays);
+  if (!copy) {
+    return null;
   }
 
   return (
-    <div
-      className="fixed top-3 z-toast flex justify-center px-3 pointer-events-none"
-      style={{ left, width: width || undefined }}
-    >
-      <div className="w-full max-w-3xl pointer-events-auto">
-        <LicenseExpiryBannerView
-          stage={stage}
-          expiresAt={expiresAt}
-          graceDaysRemaining={graceDays}
-          onDismiss={handleDismiss}
-        />
-      </div>
-    </div>
+    <AppBanner
+      variant={copy.variant}
+      title={copy.title}
+      description={copy.description}
+      dismissKey={dismissKey(stage, expiresAt)}
+    />
   );
 }
